refactor: Split client.py into multi-file module
- Moved RateLimiter to separate rate_limiter.py file
- Created auth.py for authentication and token management
- Created http.py for HTTP client and request handling
- Created cache.py for caching functionality
- Created market_data.py for market data operations
- Created trading.py for trading operations (positions, trades)
- Created base.py combining all mixins with lifecycle methods
- Maintained backward compatibility with existing imports
- Improved code organization and maintainability
