+++ src/project_x_py/client/rate_limiter.py
@@ -1,0 +1,36 @@
+"""Rate limiting for API calls."""
+
+import asyncio
+import time
+
+
+class RateLimiter:
+    """Simple async rate limiter using sliding window."""
+
+    def __init__(self, max_requests: int, window_seconds: int):
+        self.max_requests = max_requests
+        self.window_seconds = window_seconds
+        self.requests: list[float] = []
+        self._lock = asyncio.Lock()
+
+    async def acquire(self) -> None:
+        """Wait if necessary to stay within rate limits."""
+        async with self._lock:
+            now = time.time()
+            # Remove old requests outside the window
+            self.requests = [t for t in self.requests if t > now - self.window_seconds]
+
+            if len(self.requests) >= self.max_requests:
+                # Calculate wait time
+                oldest_request = self.requests[0]
+                wait_time = (oldest_request + self.window_seconds) - now
+                if wait_time > 0:
+                    await asyncio.sleep(wait_time)
+                    # Clean up again after waiting
+                    now = time.time()
+                    self.requests = [
+                        t for t in self.requests if t > now - self.window_seconds
+                    ]
+
+            # Record this request
+            self.requests.append(now)
